[Test] Migrate unit tests from entrypoints/openai/tool_parsers to tests/tool_parsers
Move 6 pure unit test files and extract unit test portions from 2 hybrid
files so that CPU-only tool parser tests live under tests/tool_parsers/
(runs in misc.yaml, no GPU) while integration tests using
RemoteOpenAIServer stay under tests/entrypoints/openai/tool_parsers/.

- git mv 6 pure unit test files (pythonic, llama3_json, llama4_pythonic,
  gigachat3, hunyuan_a13b, olmo3)
- Add function-scoped default_tokenizer override in gigachat3 tests to
  prevent tokenizer mutation from leaking across tests
- Split test_hermes_tool_parser.py: unit tests → tests/tool_parsers/,
  integration tests remain in place
- Split test_granite4_tool_parser.py: unit test → tests/tool_parsers/,
  integration test remains in place
- Delete now-unused entrypoints conftest.py
