+++ tests/tool_parsers/test_hunyuan_a13b_tool_parser.py
