+++ tests/tool_parsers/test_olmo3_tool_parser.py
